GB View changes and utils
Moved shared utils to common/utility and update genomebrowser view to handle new package updates. Also added distance field to table.
