+++ package.json
@@ -37,7 +37,7 @@
     "@visx/shape": "^3",
     "@visx/text": "^3",
     "@visx/tooltip": "^3",
-    "@weng-lab/genomebrowser": "^0.0.22",
+    "@weng-lab/genomebrowser": "^0.0.23",
     "@weng-lab/psychscreen-ui-components": "^2.0.24",
     "bpnet-ui": "^0.3.8",
     "d3-scale-chromatic": "^3.1.0",
