--- src/app/[elementType]/[elementID]/[[...tab]]/_SharedTabs/utils.ts
@@ -1,76 +1,0 @@
-import { GenomicRange } from "types/globalTypes";
-
-/**
- *
- * @param region {chrom, start, end}
- * @param transcripts
- * @returns distance to nearest TSS from any point in inputted region.
- */
-export function calcDistToTSS(
-  region: GenomicRange,
-  transcripts: { id: string; coordinates: GenomicRange }[],
-  strand: "+" | "-"
-): number {
-  const distances: number[] = transcripts.map((transcript) =>
-    calcDistRegionToPosition(
-      region.start,
-      region.end,
-      "closest",
-      strand === "+" ? transcript.coordinates.start : transcript.coordinates.end
-    )
-  );
-  return Math.min(...distances);
-}
-
-/**
- *
- * @param start Start of Region
- * @param end End of Region
- * @param anchor The anchor of region to be used: start, end, middle, or closest (finds minimum of all anchors)
- * @param point Point to Find Distance to
- * @returns The distance from the anchor specified to the position
- */
-export function calcDistRegionToPosition(
-  start: number,
-  end: number,
-  anchor: "closest" | "start" | "end" | "middle",
-  point: number
-): number {
-  const distToStart = Math.abs(start - point);
-  const distToEnd = Math.abs(end - point);
-  const distToMiddle = Math.abs((start + end) / 2 - point);
-
-  if (start <= point && point <= end) {
-    return 0;
-  }
-
-  switch (anchor) {
-    case "start":
-      return distToStart;
-    case "end":
-      return distToEnd;
-    case "middle":
-      return distToMiddle;
-    case "closest":
-      return Math.min(distToStart, distToEnd, distToMiddle);
-  }
-}
-
-/**
- *
- * @param coord1
- * @param coord2
- * @returns the smallest distance from any point in either region
- */
-export function calcDistRegionToRegion(
-  coord1: { start: number; end: number },
-  coord2: { start: number; end: number }
-): number {
-  if (coord1.end < coord2.start) {
-    return coord2.start - coord1.end;
-  } else if (coord2.end < coord1.start) {
-    return coord1.start - coord2.end;
-  } else {
-    return 0;
-  }
-}
